add file check

+++ git/setup_git.sh
@@ -1,3 +1,7 @@
-rm ~/.gitconfig
+if [ -e ~/.gitconfig ]
+then 
+    rm ~/.gitconfig
+fi
+
 ln -s "${HOME}/dotfiles/git/gitconfig" ~/.gitconfig
